pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ritsu-pao"
version = "0.1.0"
description = "因果律メディア配信パイプライン"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "jinja2>=3.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "tweepy>=4.14",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
